Added a check to the domain commands for extra positional arguments.
In much of our documentation we say the clusters in the domain commands should be specified as `--cl cluster1 cluster2` however this does not work with urfave v2. Here we have to specify it as `--cl cluster1,cluster2`.

Before this diff anything after the first positional argument was ignored, so in the command `cadence --do foo domain register --cl c1 c2 --st securityToken` 
is interpreted as `cadence --do foo domain register --cl c1 ` with `c2 --st securityToken` as ignored positional arguments. This means the error will be "No permission to do this operation." which is very misleading.